+++ src/main/java/lotto/domain/lotto/Lotto.java
@@ -1,0 +1,9 @@
+package lotto.domain.lotto;
+
+public class Lotto {
+    private final LottoNumbers numbers;
+
+    public Lotto(final LottoNumbers numbers) {
+        this.numbers = numbers;
+    }
+}
